orka gora om hela fucking spelet

+++ GrandLarceny/GrandLarceny/GrandLarceny/DevelopmentState.cs
@@ -13,8 +13,7 @@ namespace GrandLarceny
 	class DevelopmentState : States
 	{
 		#region Members
-		private LinkedList<GameObject> m_gameObjectList;
-		private LinkedList<GameObject> m_buildObjectList;
+		private LinkedList<GameObject>[] m_gameObjectList;
 		private LinkedList<GuiObject> m_guiList;
 		private LinkedList<Button> m_buttonList;
 		private LinkedList<Button> m_assetButtonList;
@@ -115,16 +114,17 @@ namespace GrandLarceny
 			m_guiList			= new LinkedList<GuiObject>();
 			m_textList			= new LinkedList<Text>();
 			m_buttonList		= new LinkedList<Button>();
-			m_buildObjectList	= new LinkedList<GameObject>();
 			m_assetButtonList	= new LinkedList<Button>();
 			m_lineList			= new LinkedList<Line>();
 			m_objectPreview		= null;
 			m_courierNew		= Game.getInstance().Content.Load<SpriteFont>("Fonts\\Courier New");
 
-			foreach (GameObject t_gameObject in m_gameObjectList) {
-				if (t_gameObject is Player) {
-					m_player = t_gameObject;
-					break;
+			foreach (LinkedList<GameObject> t_GOArr in m_gameObjectList) {
+				foreach (GameObject t_gameObject in t_GOArr) {
+					if (t_gameObject is Player) {
+						m_player = t_gameObject;
+						break;
+					}
 				}
 			}
 
@@ -251,15 +251,15 @@ namespace GrandLarceny
 		private void updateKeyboard()
 		{
 			if (keyClicked(Keys.D1))
-				m_currentLayer = 1;
+				m_currentLayer = 0;
 			if (keyClicked(Keys.D2))
-				m_currentLayer = 2;
+				m_currentLayer = 1;
 			if (keyClicked(Keys.D3))
-				m_currentLayer = 3;
+				m_currentLayer = 2;
 			if (keyClicked(Keys.D4))
-				m_currentLayer = 4;
+				m_currentLayer = 3;
 			if (keyClicked(Keys.D5))
-				m_currentLayer = 5;
+				m_currentLayer = 4;
 
 			if (keyClicked(Keys.R))
 				Game.getInstance().setState(new GameState(m_levelToLoad));
@@ -295,7 +295,7 @@ namespace GrandLarceny
 				setBuildingState(State.GuardDog);
 			if (keyClicked(Keys.Space))
 				if (m_gameObjectList != null)
-					Game.getInstance().m_camera.setPosition(m_gameObjectList.First().getPosition().getGlobalCartesianCoordinates());
+					Game.getInstance().m_camera.setPosition(m_gameObjectList[m_currentLayer].First().getPosition().getGlobalCartesianCoordinates());
 
 			if (m_currentKeyboard.IsKeyDown(Keys.LeftControl) && m_currentKeyboard.IsKeyDown(Keys.S) && m_previousKeyboard.IsKeyUp(Keys.S))
 			{
@@ -306,16 +306,18 @@ namespace GrandLarceny
 				}
 				Level t_saveLevel = new Level();
 				t_saveLevel.setLevelObjects(m_gameObjectList);
-				Serializer.getInstace().SaveLevel(m_levelToLoad, t_saveLevel);
+				Serializer.getInstance().SaveLevel(m_levelToLoad, t_saveLevel);
 
 			}
 			if (m_currentKeyboard.IsKeyDown(Keys.LeftControl) && m_currentKeyboard.IsKeyDown(Keys.O) && m_previousKeyboard.IsKeyUp(Keys.O))
 			{
-				Level t_newLevel = Serializer.getInstace().loadLevel(m_levelToLoad);
-				m_gameObjectList = t_newLevel.getLevelObjects();
-				foreach (GameObject f_gb in m_gameObjectList)
-				{
-					f_gb.loadContent();
+				Level t_newLevel = Serializer.getInstance().loadLevel(m_levelToLoad);
+				m_gameObjectList = t_newLevel.getLevelLists();
+				foreach (LinkedList<GameObject> t_arr in m_gameObjectList) {
+					foreach (GameObject f_gb in t_arr)
+					{
+						f_gb.loadContent();
+					}
 				}
 			}
 		}
@@ -420,7 +422,7 @@ namespace GrandLarceny
 			if (m_currentMouse.RightButton == ButtonState.Released && m_previousMouse.RightButton == ButtonState.Pressed) {
 				if (m_dragLine != null) {
 					if (m_selectedObject is LampSwitch) {
-						foreach (GameObject t_gameObject in m_gameObjectList) {
+						foreach (GameObject t_gameObject in m_gameObjectList[m_currentLayer]) {
 							if (t_gameObject is SpotLight) {
 								if (t_gameObject.getBox().Contains((int)m_worldMouse.X, (int)m_worldMouse.Y)) {
 									connectSpotLight((SpotLight)t_gameObject, (LampSwitch)m_selectedObject);
@@ -457,7 +459,7 @@ namespace GrandLarceny
 				}
 
 				if (!collidedWithGui(m_worldMouse)) {
-					foreach (GameObject t_gameObject in m_gameObjectList) {
+					foreach (GameObject t_gameObject in m_gameObjectList[m_currentLayer]) {
 						if (t_gameObject is LightCone)
 							continue;
 						if (t_gameObject.getBox().Contains((int)m_worldMouse.X, (int)m_worldMouse.Y)) {
@@ -466,7 +468,7 @@ namespace GrandLarceny
 							}
 						}
 					}
-					if (m_itemToCreate == State.Delete) {
+					if (m_itemToCreate == State.Delete && m_selectedObject != null) {
 						deleteObject(m_selectedObject);
 						m_selectedInfoV2 = Vector2.Zero;
 						m_selectedObject = null;
@@ -507,7 +509,7 @@ namespace GrandLarceny
 		public bool collidedWithObject(Vector2 a_coordinate) {
 			Rectangle t_rectangle = new Rectangle((int)getTile(a_coordinate).X, (int)getTile(a_coordinate).Y, 1, 1);
 
-			foreach (GameObject t_gameObject in m_gameObjectList) {
+			foreach (GameObject t_gameObject in m_gameObjectList[m_currentLayer]) {
 				if (t_gameObject is Environment)
 					continue;
 				if (t_gameObject.getBox().Contains(t_rectangle))
@@ -816,9 +818,9 @@ namespace GrandLarceny
 			if (a_gameObject is SpotLight) {
 				LightCone t_lightCone = ((SpotLight)a_gameObject).getLightCone();
 				if (t_lightCone != null)
-					m_gameObjectList.Remove(t_lightCone);
+					m_gameObjectList[m_currentLayer].Remove(t_lightCone);
 			}
-			m_gameObjectList.Remove(a_gameObject);
+			m_gameObjectList[m_currentLayer].Remove(a_gameObject);
 
 		}
 
@@ -858,7 +860,7 @@ namespace GrandLarceny
 				t_textObject.draw(a_spriteBatch);
 			foreach (GuiObject t_guiObject in m_guiList)
 				t_guiObject.draw(a_gameTime);
-			foreach (GameObject t_gameObject in m_gameObjectList)
+			foreach (GameObject t_gameObject in m_gameObjectList[m_currentLayer])
 				t_gameObject.draw(a_gameTime);
 			foreach (Button t_button in m_buttonList)
 				t_button.draw(a_gameTime, a_spriteBatch);
@@ -878,7 +880,7 @@ namespace GrandLarceny
 
 		public override void addObject(GameObject a_object)
 		{
-			m_gameObjectList.AddLast(a_object);
+			m_gameObjectList[m_currentLayer].AddLast(a_object);
 		}
 	}
 }
